requirements.txt
accelerate==0.20.3
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
azure-cognitiveservices-speech==1.30.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
decorator==4.4.2
diffusers==0.17.1
Django==4.2.2
ffmpeg-python==0.2.0
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
huggingface-hub==0.15.1
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.7.0
Jinja2==3.1.2
MarkupSafe==2.1.3
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.25.0
openai==0.27.8
packaging==23.1
Pillow==10.0.0
proglog==0.1.10
psutil==5.9.5
pydub==0.25.1
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
sqlparse==0.4.4
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.7.0
urllib3==2.0.3
yarl==1.9.2
zipp==3.15.0
